Enable netstandard1.3 and coreapp1.1
