+++ src/SpatialUnitTests/Serialization/BinaryFormatterTests.cs
@@ -1,5 +1,8 @@
 ﻿namespace MathNet.Spatial.UnitTests
 {
+
+#if NETCOREAPP1_1 == false
+
     using System.IO;
     using System.Linq;
     using System.Runtime.Serialization.Formatters.Binary;
@@ -204,4 +207,7 @@
             }
         }
     }
+
+#endif
+
 }
