+++ src/SpatialUnitTests/SpatialUnitTests.csproj
@@ -2,7 +2,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFrameworks>net45;netcoreapp2.0</TargetFrameworks>
+    <TargetFrameworks>net45;netcoreapp1.1;netcoreapp2.0</TargetFrameworks>
     <GenerateAssemblyInfo>True</GenerateAssemblyInfo>
     <AssemblyName>MathNet.Numerics.UnitTests</AssemblyName>
     <RootNamespace>MathNet.Numerics.UnitTests</RootNamespace>
@@ -27,7 +27,7 @@
     <PackageReference Include="NUnit" Version="3.9" />
     <PackageReference Include="NUnitLite" Version="3.9" />
     <PackageReference Include="protobuf-net" Version="2.3.3" />
-    <ProjectReference Include="..\Spatial\Spatial.csproj" />
+     <ProjectReference Include="..\Spatial\Spatial.csproj" />
   </ItemGroup>
   <ItemGroup>
     <Analyzer Include="..\..\packages\analyzers\StyleCop.Analyzers\analyzers\dotnet\cs\StyleCop.Analyzers.CodeFixes.dll">
@@ -41,7 +41,17 @@
       <Private>True</Private>
     </Reference>
   </ItemGroup>
-  <ItemGroup Condition="$(TargetFramework.Contains('netcoreapp2.0'))">
+  <ItemGroup Condition="'$(TargetFramework)' == 'netcoreapp1.1'">
+    <Reference Include="MathNet.Numerics">
+      <HintPath>..\..\packages\MathNet\MathNet.Numerics\lib\netstandard1.3\MathNet.Numerics.dll</HintPath>
+      <Private>True</Private>
+    </Reference>
+    <PackageReference Include="System.Runtime.Serialization.Formatters">
+      <Version>4.3.0</Version>
+    </PackageReference>
+    <PackageReference Include="System.Globalization" Version="4.3.0" />
+  </ItemGroup>
+  <ItemGroup Condition="'$(TargetFramework)' == 'netcoreapp2.0'">
     <Reference Include="MathNet.Numerics">
       <HintPath>..\..\packages\MathNet\MathNet.Numerics\lib\netstandard2.0\MathNet.Numerics.dll</HintPath>
       <Private>True</Private>
